requirements.txt
selenium~=4.8.2
pytest-rerunfailures
pytest-html-reporter